更新dotnet版本 修改弹幕解析器结构

+++ Config.cs
@@ -6,6 +6,6 @@ namespace TerminalDanmakuChan
         public string buvid3;
         public string sessdata;
         public long uid;
-        public long roomId;
+        public int roomId;
     }
 }
